backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic[email]==2.12.5
email-validator==2.1.0
httpx==0.26.0
anyio==4.12.1
starlette==0.50.0
google-generativeai==0.8.3
